refactor: extract service expander utilities to @vm0/core
Extract service expansion logic (expandServiceConfigs, validateRule,
validateBaseUrl, resolveServiceConfig, collectAndValidatePermissions)
and their tests from CLI compose command into @vm0/core contracts
module for better reusability across packages.
